requirements.txt
selenium~=4.34.2
pytest~=8.4.1
pytest-html
webdriver-manager~=4.0.2
openpyxl~=3.1.5
allure-python-commons~=2.15.0